sevalink-ai-backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
pydantic>=2.9.0
Pillow>=10.4.0
python-multipart>=0.0.12
requests>=2.32.0
google-cloud-vision>=3.7.0
google-generativeai>=0.8.0
python-dotenv>=1.0.1
alembic>=1.13.0
scikit-learn>=1.3.0
joblib>=1.3.0
scipy>=1.11.0
fastapi-jwt-auth>=0.5.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
